requirements.txt
ag
jsonref
python-dotenv
a2a-sdk==0.3.11
accelerate==1.12.0
ag-ui-protocol
agent-framework==1.0.0b251209
agent-framework-a2a==1.0.0b251209
agent-framework-ag-ui==1.0.0b251209
agent-framework-anthropic==1.0.0b251209
agent-framework-azure-ai==1.0.0b251209
agent-framework-azure-ai-search==1.0.0b251209
agent-framework-azurefunctions==1.0.0b251209
agent-framework-chatkit==1.0.0b251209
agent-framework-copilotstudio==1.0.0b251209
agent-framework-core==1.0.0b251209
agent-framework-declarative==1.0.0b251209
agent-framework-devui==1.0.0b251209
agent-framework-lab==1.0.0b251024
agent-framework-mem0==1.0.0b251209
agent-framework-purview==1.0.0b251209
agent-framework-redis==1.0.0b251209
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.72.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
asgiref==3.11.0
attrs==25.4.0
azure-ai-agents==1.2.0b5
azure-ai-evaluation==1.13.7
azure-ai-projects==2.0.0b2
azure-cognitiveservices-speech==1.47.0
azure-common==1.1.28
azure-core==1.36.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-functions==1.24.0
azure-functions-durable==1.4.0
azure-identity==1.26.0b1
azure-monitor-opentelemetry==1.8.3
azure-monitor-opentelemetry-exporter==1.0.0b46
azure-search-documents==11.7.0b2
azure-storage-blob==12.27.1
foundry-local-sdk==0.5.1
openai==2.7.2
requests>=2.31.0

# MCP Server Dependencies
fastmcp>=3.0.0b1
asyncpg>=0.29.0
pydantic>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0